provider/vsphere: organise VMs in folders
The main change in this branch is to organise
vCenter/ESXi VMs into folders. We create a
folder per controller, containing folders for
each model within that controller. The model
folders contain the VMs.

This change enables us to properly support
kill-controller, so that hosted model resources
are cleaned up even when the controller is
not available.

We also start using the standard instance tags
as VM metadata ("extra config"), and use them
for identifying controller instances.

Upgrade steps are included to move VMs into
folders, and to update VM metadata.