cmd/modelcmd: return ModelCommand from Wrap
This will allow commands that use modelcmd.Wrap
to inject the client store and the API open function
directly rather than relying on a custom NewXXXCommandForTest
initializer.

This hopefully means that we'll be able to have significantly less
ad hoc hackery to inject the standard mocked parts
into juju commands and more confidence that
ModelCommand commands aren't using the real
client store and/or API opening functions unless
we explicitly decide to.

Changes to the tests in a subsequent PR. No
externally visible behaviour changes in this one.

Credit to Ales Stimec for the idea.