provider/vsphere: refactor
Refactor the vSphere provider to use the new
internal/vsphereclient package, making it
easier to test the provider. Rather than setting
up fake vCenter folder structures, which expose
the inner workings of the VI SDK, we parameterise
the provider with a DialFunc, and pass in a mock
Client implementation so we can mock at a level
closer to the provider code.

We also stop repeating Login/Logout calls by running
each exposed method as a "session". Each session
will have one Login/Logout (with a couple of
exceptions, e.g. Bootstrap.). This enables us
to cut down on requests, and also perform some
optimisations where we cache results for the
length of a session.

There is one other functional change, which has to
do with how we configure eth1 when "external-network"
is configured. Previously we would wait for the
machine to have an IP adddress assigned, and then
SSH in and configure the iptables rules. Now we
just include them as cloud-init bootcmds.