upgrade to the latest juju/txn.
This brings in a lot of improvements to how we do live pruning.
If the number of transactions is high, it uses a database table
to prune instead of loading everything into memory. When possible,
it uses the more efficient $out aggregation pipeline to generate
that table rather than loading the values into the client.
It walks all documents to find completed transactions and remove
them from the txn-queue, rather than leaving them as referenced and
thus 'unremovable'.
It also removes items from txns.stash that have no more txn references,
keeping that collection from growing out of bounds.
The queries for finding documents that need to be touched have been
given filters to allow skipping most documents that have already
been processed. If necessary, we can start adding indexes to speed
up that query, but currently following up a prune with another prune
on a large production database only takes <200ms.