provider/vsphere: cache OVA contents on controller
When starting an instance, cache the OVA contents
on the controller's disk, and reuse if it already
exists. This can provide a substantial speedup
when starting multiple VMs.

We must use a system-wide mutex to prevent the
provisioners of separate models from trampling on
each other.

Fixes https://bugs.launchpad.net/juju/+bug/1681640