state/presence: Make beings docs smaller
There is no need to stored the model-uuid separately as it's already
in the _id prefix. The query on model-uuid can be replaced with a
prefix match on _id. This uses the index on _id speeding up the
findAllBeings query.

Given that many beings documents are created this as a significant
impact on disk space usage.

Part of the fix for https://bugs.launchpad.net/juju/+bug/1454661