provider/maas: use block device id_path
When using MAAS 1.x, we use use the block
device id_path attribute to identify its
hardware ID. If that is not available, we
fall over to "name" (i.e. the device name),
which will be set for virtio devices.

With MAAS 2.x, we have been using the "path"
attribute, which turns out to be set with
paths that do not really exist, or perhaps
depend on the kernel version. Anyway, the
MAAS docs say that id_path should not change
depending on the kernel version, so we will
now use that with both MAAS 1.x and 2.x.

Fixes https://bugs.launchpad.net/juju/+bug/1677001