provider/vsphere: various fixes
There are a few fixes bundled up here, for
issues found while testing against a resource
limited vCenter 6.5 installation.

The first and main change is to update the
lease obtained while uploading the VM image
and metadata. If this is not done, the
lease may expire, which will cause the VM
creation to fail. As part of making this
change, we now update the progress periodically
rather than when meeting/exceeding a bytes
transferred delta. This provides better
feedback in slow environments, and also
ensures the lease is updated regularly.

The second change is to not provide defaults
for cpu-cores, cpu-power, or mem. These
constraints not be satisfiable by all vSphere
environments.

The thid and final change is to stop making a
PowerOffVM_Task call for VMs that are not
powered on, when removing VMs. Making that
call causes an error, which prevents VM
removal.

Fixes https://bugs.launchpad.net/juju/+bug/1682393