cmd/modelcmd: remove SetAPIOpener
We have both SetAPIOpen and SetAPIOpener which are mutually redundant
and confusing.  SetAPIOpener is barely used, so remove it and mock
api.Open instead, which means that more logic is tested.

Also move the NewTimeoutOpener logic into the kill command, the
only place that uses it, so we avoid using the APIOpener abstraction
in that case. At some point in the future, it would be nice to make
juju.NewAPIConnection cancellable (for example by passing
in a context.Context) which would make the connection timeout
logic more generally available.