Decouple GCE firewall name from source CIDRs
We can't update the firewall name when we change the source CIDRs in it,
because GCE rejects the update (since the name doesn't match the URL).
That means we can't use hashes of the CIDRs as the name - when we change the
CIDRs then we'll no longer be able to find the matching rule since it
won't have the name we expect based on the CIDRs.

Instead, assign random names to the firewall rules as they're created,
and ensure we preserve them in the RuleSet so that we can send them back
to GCE when updating an existing rule.

We still use hashing of the CIDRs to match input rules against existing
rules to determine whether to update or create a rule.

UNTESTED