cmd/juju/application: make deploy init tests assume less about internals
The tests are doing a deep-equal comparison of the deploy
command, which assumes more than it should about what
the state of the embedded objects. Make the code check
the fields it cares about explicitly.

Also clean up things a little bit:
- NewDeployCommand is now the standard constructor, consistent
with the other commands.
- use runDeploy throughout, losing the hard-to-remember distinction
between runDeploy and runDeployCommand.