cmd/modelcmd: rename CommandBase to Command
The "CommandBase" name is inconsistent with all the other *Base
names - it's an interface but all the others (ControllerCommandBase,
ModelCommandBase and cmd.CommandBase) are static embeddable structs.
This is confusing, so rename it to Command, allowing us to rename the
awkwardly-differentiated JujuCommandBase to CommandBase.

This PR was entirely mechanically generated with

	gorename -from '"github.com/juju/juju/cmd/modelcmd".CommandBase' -to Command
	gorename -from '"github.com/juju/juju/cmd/modelcmd".JujuCommandBase' -to CommandBase

with the exception of some of the comment adjustments.