apiserver/remotefirewaller: Fix test data races
Rather than creating the mock machine address watcher lazily, create it
when the machine is created. This prevents a data race where the code
under test hasn't asked for the watcher (and thus created it) when we
try to send a change.

Put a lock around reading and writing mockUnit.publicAddress.