state: volume-backed filesystems have volume scope
Volume-backed filesystems now take on the scope of
the backing volume. If the volume is model-scoped,
then so is the filesystem. This enables us to detach
and reattach volume-backed filesystem storage.

The watcher implementation for machine filesystems
and filesystem attachments has been updated to
report filesystems whose backing volumes are attached
to the specified machine.