Capture space information from the remote provider
...when a remote application is created.

Rearranged application API remote application code to make more use of
ApplicationOffer - now that we can rely on an offer existing we don't
need to synthesise one from the application. Use offer endpoints rather
than the application's (ensures we only create remote endpoints for the
offered ones).

Rather than passing the source model tag around separately, use the one
from the params.ApplicationOffer.

Creating the params.ApplicationOffer inside
sameControllerOfferedApplication simplifies the logic for ensuring that
the offering model gets released to the pool.

We pass space info into the provider because we can't rely on the
provider to know which subnets are in a space - MAAS is the only
provider which does. Passing the space info from the state database in
allows the other providers to return sensible ProviderSpaceInfos
including the provider-specific information they have.

Use environs.DefaultSpaceName in EndpointBindings and define a local
defaultEndpointName constant for the places using empty string as a
binding key.