all: remove ExternalControl from migration
Following from a remark from Menno, both ExternalControl
and SkipInitialPrechecks were only there if full migrations
were not implemented in time, but they're implemented now
so we can remove them.

This prepares for adding CACert-fetching logic to the
initial pre-check in the apiserver InitiateMigration logic - that
didn't make sense if the pre-check was optional.