Update the cross-model methods in NetworkingEnviron
After discussing the interface with jam it became clear that the
original definitions of the methods weren't quite right.

With the exception of MAAS which fully supports spaces, providers don't
have any way to determine which subnets should be in a space. State is
the authority in that case, so just passing ProviderId into
ProviderSpaceInfo can't work. (We could change the providers to tag
subnets with space information, but that's not straightforward - at the
provider level the subnets are shared between controllers, so multiple
controllers trying to manage the tags on one subnet could clobber each
other's changes unless care was taken.)

Change ProviderSpaceInfo to take a network.SpaceInfo with state's view
of the subnets instead - if the provider knows better it can discard
that, but otherwise the space info can be included in the result.

Also rename IsSpaceRoutable to AreSpacesRoutable, which takes two
spaces. The arguments to this will be the ProviderSpaceInfos for the two
endpoints of the cross-model relation.