Update juju/txn
This allows us to pass in a window of recent transactions to
preserve. As a reasonable balance of being conservative vs
keeping the transaction queue small, we preserve the last
hour of transactions. This matches the frequency that we will
be checking to see if there are transactions that need to be pruned.