Don't query remoteApplications without CMR flag
Controllers bootstrapped without the crossmodel feature flag enabled
would occasionally log `using unknown collection
"remoteApplications". This is because the remoteApplications collection
isn't created in that case, and state/applicationByName was querying it.

Update applicationByName to check st.Application first, and only fall back
to trying st.RemoteApplication if the flag is on.