Don't track relations irrelevant to subordinate units
Make WatchApplicationRelations take units. Renamed it to
WatchUnitRelations, and updated the client facade and the uniter worker
to use it there.

If the requesting unit is a subordinate, we filter out any
container-scoped relations that aren't to the principal's
application. This prevents these relations from keeping a subordinate
unit alive if the relation between the subordinate application and its
principal's application is removed.

The uniter should still see other global-scoped relations these - the
existence of a global-scoped relation won't keep a subordinate unit
alive.

Rationalised the different versions of the API - there had been a few
confusing changes that left us in a state that didn't make much
sense. The result of the change is that we still have v4 and v5, with
the following differences:

v4 has the now-obsolete methods NetworkConfig and
WatchApplicationRelations.

v5 has the new methods SLALevel, NetworkInfo and WatchUnitRelations.

@wupeka pointed out that the reflection machinery ignores 2-argument
methods, so we can prevent a method from being exposed in the API by
hiding it behind a 2-argument one. This enables a simpler structure for
different API versions.