Add volume attachments to StartInstanceParams
Add VolumeAttachments to the StartInstanceParams
struct, which contains the params for attaching
existing volumes to the instance being created.
The provisioner is not required to attach the
volume; it will still be attached by the storage
provisioner as we do today. The new field is
added so that volume attachments may influence
how instances are created, e.g. by constraining
the instance to a specific availability zone.