Fix race in two apiserver watchers calling EnsureErr
In both subRelationsWatcher and IngressAddressWatcher the use of the
watcher.EnsureErr was causing a potential race.

If the loop saw a closed Changes channel from the child
watcher (relationsw in the subRelationsWatcher, ruw in the ingress
address watcher) because the catacomb had stopped it, then w.Err() would
be nil in EnsureErr. Formatting the watcher as part of the error message
would then trigger a data race since it would examine mutex data inside
the tomb.

We don't actually need to worry about collecting the error that killed
the child watcher here (if it's not just dying because it was asked),
since the catacomb will propagate the error itself. So just return
w.catacomb.ErrDying() instead.

An example of the race can be seen here:
http://reports.vapour.ws/releases/5275/job/run-unit-tests-race/attempt/2764