juju: use new DNS caching functionality
We rely on the new API client functionality to do our
DNS resolving for us, and store any resulting cached
names in the controllers.yaml entry.

The UnresolvedAPIEndpoints field is now redundant
and is replaced by the strictly advisory DNSCache
field. The APIEndpoints field now always contains all
the host names returned from the Login call with
the exception of addresses deemed unusable
(e.g. link-local addresses).

We also change rpcreflect.Value.FindMethod to accept any
version so that it's more useful for testing - this makes it
possible to server limited portions of the juju API from
a mocked API type without implementing a custom FindMethod.
And, in passing, fix rpc.NewConn so that it works with a nil obververFactory
as documented.

QA check that API connections still work OK, particularly on
MAAS setups where some of the returned host names will
not resolve on a client.

Also, check that:

	juju register jaas
	juju list-models

does not produce the "cannot validate certificate for because it
doesn't contain any IP SANs" error.

Fixes https://bugs.launchpad.net/juju/+bug/1692905