provider/vsphere: truncate folder name to 80 chars
There is an 80 character limit on folder names in
vSphere, so we truncate the model name. The model
name limit is 33 characters, which should be more
than enough for most model names. Truncating just
ensures that we do not fail in pathological cases.

Fixes https://bugs.launchpad.net/juju/+bug/1691347