Update log forwarder to work per-model
Move it into the models manifolds. This means we can drop the fake state
dependency - this wasn't required except to prevent the worker from
running in machine agents that weren't the controller, and only the
controller runs model workers.

Pull out the support for forwarding logs from all models. We want to
split the logs collection into one collection per model after this
change, so logging from all models would be harder to do (and we don't
think there's any need for it).

Make the forwarder always send the model ID.