Fix potential state leak in machine agent
...and make the syslog feature test pass.The test was failing for a
couple of weird reasons (despite the feature working correctly).

First, the log-forwarder was never being started so the dummy rsyslog
server never got any messages. This was because the model manifolds are
run by the state workers worker, which in turn depended on the state
worker (which opens a state connection). This was failing because the
default controller config used when creating the agent config had the
state/mongo port as 1234. Explicitly setting this to the port reported
by the juju/testing mongo instance resolved this, revealing the next
problem.

The test was now passing (log forwarding was working) but teardown was
failing because there were mongo connections left alive. This turned out
to be because there was a state leak in the machine agent - if there was
an error starting the apiserver, the state created for it would be
leaked. The apiserver wasn't starting because it was trying to bind to a
port that was in use. It turns out that the dummy provider starts an
apiserver on that same port, and all of the workers that used the API
are talking to that, instead of the one that the agent should've
created.

I've fixed the leak, and the test passes, but it's a weird situation. I
think the right thing would be for tests against a model-hosting
agent (ie a controller agent rather than a normal machine or unit agent)
not to use AgentSuite, but to use something else that doesn't use dummy
provider to bootstrap, since the tests will run an agent that should
contain its own apiserver.

I'll make a separate task for that.