api: fix TestFallbackToIPLookupWhenCacheOutOfDate flakiness
In TestFallbackToIPLookupWhenCacheOutOfDate, we were expecting both
dial attempts to have been called before the api.Open call returned.
However, each dial attempt is called in its own goroutine, so the second
address can start and complete before the first one starts, at which
point the attempt is cancelled and the first address retry loop will
never be entered.

We fix this by waiting synchronously for the dial
attempts and only letting them complete
when we see that they've happened.