debug/server.go: change debugging session tips
Currently it is not mentioned how to exit a debugging session and
prevent any further manual event processing, e.g. when you are in
a wildcard debug-hooks session you might be getting new events too fast
and a simple shell exit in an event processing window will result in an
immediate processing of a new event from the backlog.