worker/introspection: improve shell functions
This branch brings several improvements to the
agent introspection shell functions:

- we no longer use socat to communicate with the
  introspection socket, which was causing problems
  with the CPU profile endpoint. Instead, we add
  a new "juju-introspect" command/symlink to jujud,
  which serves as a frontend to the introspection
  socket. The helpers for getting specific profiles
  are left as bash functions.
- the new juju-introspect command has a --listen
  flag which, if specified, causes the command to
  expose the socket over HTTP. This enables the
  user to use "go tool pprof" with the HTTP endpoint,
  for example.
- A new function, juju-cpu-profile, is added, which
  will capture a CPU profile. The first (optional)
  argument is interpreted as the number of seconds
  to sample for.

Since we are no longer writing out the raw HTTP output,
the output of the various juju-foo-profile functions
no longer include the HTTP response header, so the
output can be fed directly into "go tool pprof".

By moving the main frontend out of the bash script and
into Go code, we get a step closer to having
introspection on Windows. We still need an alternative
to the abstract domain sockets solution that we have
today (probably named pipes?)