Fix deleting status history in Unit.Destroy
The deletion code expected the key to be in the `statusid` field, but
the status history docs were changed to have the global key stored as
`globalkey`, so the `RemoveAll` calls were never deleting anything and
always doing a full scan.

Add a test that the status history is gone and correct the
deletion. Also delete workload version status records. Move
eraseStatusHistory to status - keeping everything that needs to know
about the fields of the underlying collection in one place reduces the
likelihood of them getting out of sync.

Fixes https://bugs.launchpad.net/juju/+bug/1696491