api: impose stricter deadline on api.Open
Currently if you call api.Open with a given timeout,
it might exceed that timeout if the connect succeeds
but (for whatever reason) the Login API call doesn't return.

This PR imposes the timeout on the login too, so clients
that care can be more sure that the api.Open will actually
return soon after a timeout.

To avoid changing the useful semantic (relied on by many tests)
that the zero DialOpts will log in without any particular timeout,
we change the semantics slightly, such that a zero timeout
means no timeout and a zero RetryDelay means "don't retry",
which means that the zero DialOpts means pretty much what
it currently means.