apiserver: don't close state pool
apiserver.NewServer now takes a StatePool as a primary
argument, rather than a State. The Server does not take
ownership of the StatePool; it is the responsibility of
the caller to close it after the apiserver closes.

The Server still calls KillWorkers on the StatePool,
so when the apiserver closes it *must* be closed and
a new State and StatePool opened for a new apiserver
to be started. Nevertheless, this change enables us
to use the StatePool in the statemetrics code, and gets
us a step closer to putting the apiserver into the
dependency engine as a first-class manifold.