Test that we do flush after some time.
This test is potentially flaky, but I should have made it safe in the
face of races. We *could* try to introduce clocks, but I think this is
still correct.  We make sure that we wait at least long enough, and then
continue to wait for significantly longer to make sure even under load
the flush will have succeeded.

Also, we don't have to track ModelUUID on the ping slots, because that
is already encoding in the DocID that we use in the map.