api: check before using pre-resolved address
When dialing a websocket, check that the address
we're asked to dial is the same as the one that
we pre-resolved. It will not match when we're
using a proxy.

This is a short-term fix. If we should be doing
any client-side resolution at all, it should be
disabled when a proxy is to be used.

Fixes https://bugs.launchpad.net/juju/+bug/1698989