Upgrade step to correct relation unit counts
The subordinate unit handling fix in https://pad.lv/1686696 meant that
fewer units would enter the scope of a container-scoped relation. If a
model with subordinates related to more than one application was
upgraded from the 2.1 version to 2.2, there would be records in
relationscopes for units that would never leave scope (since they were
no longer considered in scope for that relation). This meant that the
unit count of the relation would never reach 0, which in turn would
prevent the application from being removed.

Add an upgrade step to detect relationscopes rows that shouldn't be
present and remove them, updating the relation unit count to keep it
consistent.

Fixes https://bugs.launchpad.net/juju/+bug/1699050