state: Prevent RelationUnits with bad scopes
It was possible to create a RU for a subordinate unit where the
principal isn't a member of the relation. For example, if you had
logging related to mysql and wordpress, you could create an RU for a
logging unit attached to a wordpress unit in the mysql-logging relation,
and then cause havoc.

Prevent this by checking the principal is in the endpoints and raise a
specific InvalidPrincipal error - we'll want to catch and ignore this
error in the apiserver to avoid crashing agents in an upgrade.