apiserver/uniter: Ignore EnterScope with invalid principals
If an uniter for a subordinate unit requests to EnterScope for a
relation that isn't valid because its principal isn't a member of the
relation, we ignore it. This is needed for clean upgrading from 2.1 to
2.2.1 - there's an upgrade step that corrects unitcounts and invalid
relationscope records, but without this change the agents will re-enter
the scopes incorrectly when they reconnect to the controller (because we
don't get a chance to upgrade them first).