apiserver/uniter: Ignore EnterScope from invalid units
If we see a request to EnterScope where the unit's principal isn't in
the relation specified, we log and ignore it. These requests can happen
when a controller is upgraded from 2.1 to 2.2 - the uniter used to think
that those units should be in the relation, but they shouldn't, and
allowing them to enter scope will prevent the application from being
removed because its relations don't go away (since they have units in
scope that never leave).

Ignore the requests to prevent the old uniter from crashing with an
error, which would prevent upgrading them.