worker/uniter: handle storage IDs with hyphens
Some code in the uniter's storage resolver was
based on old rules for storage ID names, which
disallowed hyphens. We would translate the "/"
in a storage ID to "-" when writing to disk,
and then translating all "-" to "/" on the way
back in, and ignoring the file because it's an
invalid storage ID. Instead we should only
translate the final "-" to "/", since the name
may now contain hyphens.

Fixes https://bugs.launchpad.net/juju/+bug/1674148