Use testclock.WaitAdvance in TestPause
The test passes fine locally but fails often in CI, and when run under
stress it fails after 1 or 2 runs. It looks like the clock advance is
happening before the throttlingListener manages to wait for it. Change
the Advance calls to WaitAdvance so we know the listener is waiting when
we advance.

Removed start channel - it complicated the test for no real benefit.