Fix an assert that wasn't testing anything.
In one of the recent refactorings, I pulled out a helper
function to test whether an entity was alive. I turned out
that the helper was broken in many ways.
It hard coded the 'key', but that didn't end up mattering because
it ended up doing:
   c.Assert(alive, gc.Equals, alive)
which will always be true (it reused a parameter as a local variable).

The test happened to be correct anyway, this just fixes it to actually be
testing what we thought it was.