Possible fix for 1703526.
Instead of fixing a PingBatcher when initializing a Pinger,
we instead have the pinger ask for whatever the current PingBatcher is.
That way if the PingBatcher gets restarted due to an error, then the
Pingers can find the correct one.

Only other problem is we probably don't restart the Pingers either if
they encounter an error, so that needs testing as well.

Also, we're missing a test for the fact that Pinger's always use the
fresh PingBatcher (they do by inspection, but it should be actually tested.)