storage: add {Filesystem,Volume}Importer
Add interfaces for importing filesystems and
volumes into the model. This enables a user
to import storage into the model/controller
that was created elsewhere, e.g. by another
model/controller. Importing the storage brings
it under the control of this model/controller;
destroying the model/controller is expected
to destroy the imported storage.

This is initially implemented by LXD and EBS
only. Other providers will follow.