cmd/juju/storage: add import-filesystem command
Add the api/storage side of the Storage.Import API,
and introduce the "import-filesystem" command on
top of that. This command can be used to import an
existing filesystem into the model. The filesystem
is expected to have been provisioned already, and
will start out in the model as provisioned, having
the status "detached".

After running import-filesystem, destroying the
model or controller is expected to destroy the
imported storage.