apiserver/facades/...: support releasing storage
Update the Storage API facade to handle the new
ReleaseStorage field in the DestroyStorage method's
parameters. This is to support a client requesting
that storage be non-destructively removed from the
model.

Update the StorageProvisioner API with new methods
for obtaining parameters to destroy volumes and
filesystems (distinct from parameters for creating),
which include a new field indicating whether or
not the storage should be released rather than
destroyed.