remove-storage: add --no-destroy flag
Update the remove-storage command with a --no-destroy
flag, which instructs Juju to non-descructively remove
(AKA "release") the storage. This means the cloud storage
will no longer be under the control of Juju, but it is
not destroyed, and will not be destroyed by destroy-model
or destroy-controller.

The Storage API client is updated with an option to
non-destructively remove (aka "release") storage to
support this. Drive-by renaming of DestroyAttached to
DestroyAttachments (no need to rev, the new facade version
hasn't been released yet).