provider/vsphere: add network configuration
Add a "network" config attribute to the vsphere
provider. When bootstrapping or adding a model,
the user can set this to the name of the primary
network that VMs will be connected to. This will
override the default "VM Network" specified by
the Ubuntu images' OVF files.

Fixes https://bugs.launchpad.net/bugs/1619812