resources: Don't schedule multiple cleanups for one path
This was the cause of the failed cleanups for resources that were
blocking migrations. Unit and applications resources were both found and
cleanups were scheduled for them even if they had the same
storage-path. This meant that the first cleanup would succeed and the
second would fail.

Change ResourcePersistence.NewRemoveResourcesOps to only emit one
cleanup for each path.