state: introduce DestroyModelParams
We consolidate Model.Destroy and Model.DestroyIncludingHosted
into a single method, Model.Destroy, and introduce the
DestoryModelParams struct to control its behaviour. There are
two fields for now:

 - DestroyHostedModels. If this is true, then the behaviour
   is the same as the removed Model.DestroyIncludingHosted
   method; otherwise the behaviour is the same as the replaced
   Model.Destroy method.
 - DestroyStorage. If this is nil, and there is any persistent
   storage in the model, then destroying the model will fail
   with an error that satisfies state.IsHasPersistentStorageError.
   If the field's value is true, then all storage will be
   destroyed. If the field's value is false, then the storage
   will be released/disassociated from the model/controller.

To implement these changes, we have to extend the cleanup doc
with cleanup-specific arguments.

To check whether storage is persistent in hosted models, we
have had to change various volume and filesystem query functions
to operate on a database, instead of a state object. This
avoids an expensive State.ForModel call.