state: Remove state.ControllerModel()
This is part of a series of changes to clean up how Model instances
are created. The ControllerModel method has been replaced with
ControlModelUUID and ControllerModelTag methods. To get a
controller Model you must now go via a State for the controller.

To simplify one common use case, a new ControllerOwner method has been
added to State.