apiserver/facades: Remove AllModels and ForModels usage
AllModels is now replaced with AllModelUUIDs combined with
GetModel. Exposing an AllModels backend interface was problematic in
terms releasing States. Separating this out means we're not using
State instances for Models which have already been released.

Various uses for ForModel were also removed as part of this. This is
fixes https://bugs.launchpad.net/juju/+bug/1698702.