Refactor EnvironConfigGetter.CloudSpec()
This surprisingly large change changes the
environs.EnvironConfigGetter interface so the CloudSpec method no
longer takes a model tag. This means that a single instance of an
EnvironConfigGetter is now only ever responsible for a single model.

This change is required because it supports changes around state.Model
creation. A CloudSpec method that takes a model tag encourages poor
practices in EnvironConfigGetter implementations.