Remove pending resources when Deploy fails
The deploy command creates pending resources for an application before
creating the application, and links them to the application using
pending IDs. These are used to resolve the resources during application
creation. If the application creation fails for some reason, these
pending resources need to be cleaned up.

This cleanup was initially done at the state level, but manual testing
showed that if a deployment failed due to validation at the API level,
the pending resources would still be left. Doing the cleanup in the
application API instead prevents this.

Part of the fix for https://bugs.launchpad.net/juju/+bug/1705730